Fixed bug in BinomialRand generator + some refactoring

+++ math/NumericMath.cpp
@@ -17,8 +17,8 @@ namespace RandMath
  * @param bottom
  * @return
  */
-long double adaptiveSimpsonsAux(const std::function<double (double)> &funPtr, double a, double b,
-                                          double epsilon, double S, double fa, double fb, double fc, int bottom)
+double adaptiveSimpsonsAux(const std::function<double (double)> &funPtr, double a, double b,
+                           double epsilon, double S, double fa, double fb, double fc, int bottom)
 {
     double c = .5 * (a + b), h = (b - a) / 12.0;
     double d = .5 * (a + c), e = .5 * (c + b);
@@ -35,7 +35,7 @@ long double adaptiveSimpsonsAux(const std::function<double (double)> &funPtr, do
            adaptiveSimpsonsAux(funPtr, c, b, epsilon, Sright, fc, fb, fe, bottom);
 }
 
-long double integral(const std::function<double (double)> &funPtr, double a, double b, double epsilon, int maxRecursionDepth)
+double integral(const std::function<double (double)> &funPtr, double a, double b, double epsilon, int maxRecursionDepth)
 {
     // TODO: redo to adaptive Gauss-Kronrod quadrature
     if (a > b)
